requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
python-dotenv
pypdf
python-docx
langchain
langchain-text-splitters
langchain-google-genai
google-generativeai
chromadb
langchain-chroma
passlib[bcrypt]
python-jose
